feat(boards): implement board renaming (#25)
* feat(frontend): add updateBoard method to BoardService

Updates the BoardService to include a method for renaming a board via a PUT request to the /api/boards/{id} endpoint and reorganized methods.

* feat(api): implement update board endpoint

Adds a PUT /api/boards/{id} endpoint, creates BoardUpdateDto, and updates the repository to handle board renaming.

* feat(boards): implement board renaming

Adds a click-to-edit feature for the board title in BoardDetailComponent, allowing users to rename their boards.

* feat(boards): improve board detail UI and update interaction logic

Adds a validator to the titleControl and attribute to the input field so that user can`t write more than 50 symbols for the board title